feat: 完成核心功能开发和代码优化
## 新增功能

### 热点数据源扩展
- 新增抖音热点数据源 (douyin.go)
- 优化微博、知乎数据源,添加真实抓取框架
- 完善数据源接口实现 (ID, DisplayName方法)
- 支持多数据源注册 (8个数据源)

### 定时发布功能
- 实现定时任务调度系统 (scheduler.go)
- 创建调度器API处理器 (scheduler_handlers.go)
- 支持任务的创建、更新、删除、暂停、恢复、立即执行
- 添加调度器统计API

### 前端功能
- 新增定时发布页面 (ScheduledPublish.tsx)
- 创建Switch UI组件 (switch.tsx)
- 修复文件上传字段问题 (path/storage_path兼容)
- 添加上传进度回调支持

### 文档和规划
- 创建数据分析规划文档 (analytics_plan.md)
- 创建项目文件管理报告 (PROJECT_FILE_MANAGEMENT.md)
- 更新TODO.md,标记已完成的任务
- 创建项目总结文档 (PROJECT_SUMMARY.md)

## 代码优化

### 存储API完善
- 实现文件下载 (downloadFile)
- 实现文件列表 (listFiles)
- 实现文件删除 (deleteFile)
- 添加批量操作支持

### 发布流程优化
- 完善任务处理器调用实际适配器
- 添加任务状态更新机制
- 完善错误处理和结果返回

### 登出功能完善
- 实现后端登出API
- 前端调用登出API
- 清除登录状态

## 代码审核

发现并记录23个代码质量问题:
- 高严重度问题: 5个
- 中严重度问题: 12个
- 低严重度问题: 6个

主要问题包括:
- CORS配置过于宽松
- JWT密钥使用随机生成
- 错误处理不充分
- 请求超时goroutine泄漏
- 缺乏输入验证

## 项目文件管理

- 梳理项目目录结构
- 优化文档组织
- 统一文件命名规范
- 完善文件权限管理

+++ publisher-core/hotspot/monitor.go
@@ -431,7 +431,7 @@ func (s *CompetitorAnalysisService) identifyOpportunities(topic database.Topic, 
 	}
 
 	// 检查热度差距
-	if len(competitors) > 0 && topic.Heat > competitors[0].Heat*0.8 {
+	if len(competitors) > 0 && topic.Heat > int(float64(competitors[0].Heat)*0.8) {
 		opportunities = append(opportunities, "与领先者差距较小，有机会超越")
 	}
 
@@ -444,7 +444,7 @@ func (s *CompetitorAnalysisService) identifyThreats(topic database.Topic, compet
 
 	// 检查是否有强劲的上升竞品
 	for _, comp := range competitors {
-		if comp.Trend == "up" && comp.Heat > topic.Heat*0.9 {
+		if comp.Trend == "up" && comp.Heat > int(float64(topic.Heat)*0.9) {
 			threats = append(threats, fmt.Sprintf("竞品 '%s' 正在快速上升", comp.Title))
 		}
 	}
